\name{rcmudict-package}
\alias{rcmudict-package}
\alias{rcmudict}
\docType{package}
\title{
What the package does (short line)
CMU Phonetic Dictionary
}
\description{
CMU Phonetic dictionary
}
\details{
\tabular{ll}{
Package: \tab rcmudict\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-03\cr
License: \tab MIT\cr
}

}
\author{
Kurt Harriger

Maintainer: Kurt Harriger <kurtharriger@gmail.com>
}
\references{
http://www.speech.cs.cmu.edu/cgi-bin/cmudict
}
\keyword{ cmu phonetic pronouncing dictionary }
\seealso{
}
\examples{
}
